+++ source/renderer/app/components/status/DaedalusDiagnostics.js
@@ -1,10 +1,11 @@
 // @flow
 import React, { Component } from 'react';
 import { observer } from 'mobx-react';
-import { get, includes, upperFirst } from 'lodash';
+import { get, includes, upperFirst, capitalize } from 'lodash';
 import { defineMessages, intlShape } from 'react-intl';
 import moment from 'moment';
 import classNames from 'classnames';
+import CopyToClipboard from 'react-copy-to-clipboard';
 import { Tooltip } from 'react-polymorph/lib/components/Tooltip';
 import { TooltipSkin } from 'react-polymorph/lib/skins/simple/TooltipSkin';
 import SVGInline from 'react-svg-inline';
@@ -15,6 +16,7 @@ import {
 import { UNSYNCED_BLOCKS_ALLOWED } from '../../config/numbersConfig';
 import { getNetworkEkgUrl } from '../../utils/network';
 import closeCross from '../../assets/images/close-cross.inline.svg';
+import iconCopy from '../../assets/images/clipboard-ic.inline.svg';
 import externalLinkIcon from '../../assets/images/link-ic.inline.svg';
 import LocalizableError from '../../i18n/LocalizableError';
 import { CardanoNodeStates } from '../../../../common/types/cardano-node.types';
@@ -105,6 +107,16 @@ const messages = defineMessages({
     id: 'daedalus.diagnostics.dialog.stateDirectory',
     defaultMessage: '!!!Daedalus State Directory',
     description: 'Daedalus State Directory',
+  },
+  stateDirectoryPathOpenBtn: {
+    id: 'daedalus.diagnostics.dialog.stateDirectoryPathOpenBtn',
+    defaultMessage: '!!!Open',
+    description: 'Open',
+  },
+  stateDirectoryCopyPathBtn: {
+    id: 'daedalus.diagnostics.dialog.stateDirectoryCopyPathBtn',
+    defaultMessage: '!!!Copy',
+    description: 'Copy',
   },
   connectionError: {
     id: 'daedalus.diagnostics.dialog.connectionError',
@@ -333,10 +345,13 @@ type Props = {
   latestNetworkBlockTimestamp: number,
   localBlockHeight: number,
   networkBlockHeight: number,
+  currentLocale: string,
   onForceCheckLocalTimeDifference: Function,
+  onOpenStateDirectory: Function,
   onOpenExternalLink: Function,
   onRestartNode: Function,
   onClose: Function,
+  onCopyStateDirectoryPath: Function,
 };
 
 type State = {
@@ -468,7 +483,9 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
       latestLocalBlockTimestamp,
       latestNetworkBlockTimestamp,
       onForceCheckLocalTimeDifference,
+      onOpenStateDirectory,
       onClose,
+      onCopyStateDirectoryPath,
       nodeConnectionError,
       isSystemTimeIgnored,
       onOpenExternalLink,
@@ -476,6 +493,7 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
       isTestnet,
       isStaging,
       isMainnet,
+      currentLocale,
     } = this.props;
 
     const {
@@ -549,6 +567,11 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
           isTestnet,
         });
 
+    const stateDirectoryPathStyles = classNames([
+      styles.stateDirectoryPath,
+      styles[`locale-${currentLocale}`],
+    ]);
+
     return (
       <div className={styles.component}>
         <div className={styles.tables}>
@@ -561,27 +584,29 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                 </th>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.platform)}:</td>
+                <th>{intl.formatMessage(messages.platform)}:</th>
                 <td>{platform}</td>
               </tr>
               <tr className={styles.platformVersion}>
-                <td>{intl.formatMessage(messages.platformVersion)}:</td>
+                <th>{intl.formatMessage(messages.platformVersion)}:</th>
                 <td className={styles.platform}>{platformVersion}</td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.cpu)}:</td>
+                <th>{intl.formatMessage(messages.cpu)}:</th>
                 <Tooltip skin={TooltipSkin} tip={cpu}>
                   <td>{cpu}</td>
                 </Tooltip>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.ram)}:</td>
+                <th>{intl.formatMessage(messages.ram)}:</th>
                 <td>{ram}</td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.availableDiskSpace)}:</td>
+                <th>{intl.formatMessage(messages.availableDiskSpace)}:</th>
                 <td>{availableDiskSpace}</td>
               </tr>
+            </tbody>
+            <tbody>
               <tr>
                 <th colSpan={2}>
                   {intl.formatMessage(messages.coreInfo)}
@@ -589,19 +614,19 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                 </th>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.daedalusVersion)}:</td>
+                <th>{intl.formatMessage(messages.daedalusVersion)}:</th>
                 <td>{daedalusVersion}</td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.daedalusMainProcessID)}:</td>
+                <th>{intl.formatMessage(messages.daedalusMainProcessID)}:</th>
                 <td>{daedalusMainProcessID}</td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.daedalusProcessID)}:</td>
+                <th>{intl.formatMessage(messages.daedalusProcessID)}:</th>
                 <td>{daedalusProcessID}</td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.safeMode)}:</td>
+                <th>{intl.formatMessage(messages.safeMode)}:</th>
                 <td className={styles.safeMode}>
                   {isInSafeMode
                     ? intl.formatMessage(messages.statusOn)
@@ -609,28 +634,51 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                 </td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.cardanoVersion)}:</td>
+                <th>{intl.formatMessage(messages.cardanoVersion)}:</th>
                 <td>{cardanoVersion}</td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.cardanoProcessID)}:</td>
+                <th>{intl.formatMessage(messages.cardanoProcessID)}:</th>
                 <td>{cardanoProcessID}</td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.cardanoApiPort)}:</td>
+                <th>{intl.formatMessage(messages.cardanoApiPort)}:</th>
                 <td>{cardanoAPIPort || '-'}</td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.cardanoNetwork)}:</td>
-                <td>{cardanoNetwork}</td>
+                <th>{intl.formatMessage(messages.cardanoNetwork)}:</th>
+                <td>{capitalize(cardanoNetwork)}</td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.stateDirectoryPath)}:</td>
-                <Tooltip skin={TooltipSkin} tip={daedalusStateDirectoryPath}>
-                  <td className={styles.stateDirectoryPath}>
-                    {daedalusStateDirectoryPath}
-                  </td>
-                </Tooltip>
+                <th>{intl.formatMessage(messages.stateDirectoryPath)}:</th>
+                <td className={styles.stateDirectory}>
+                  <button
+                    className={styles.stateDirectoryOpenBtn}
+                    onClick={() =>
+                      onOpenStateDirectory(daedalusStateDirectoryPath)
+                    }
+                  >
+                    {intl.formatMessage(messages.stateDirectoryPathOpenBtn)}
+                  </button>
+                  <CopyToClipboard
+                    text={daedalusStateDirectoryPath}
+                    onCopy={onCopyStateDirectoryPath}
+                  >
+                    <div className={stateDirectoryPathStyles}>
+                      <Tooltip
+                        skin={TooltipSkin}
+                        tip={
+                          <div style={{ textAlign: 'center' }}>
+                            <div>{daedalusStateDirectoryPath}</div>
+                          </div>
+                        }
+                      >
+                        <p>{daedalusStateDirectoryPath}</p>
+                        <SVGInline svg={iconCopy} />
+                      </Tooltip>
+                    </div>
+                  </CopyToClipboard>
+                </td>
               </tr>
               {!isConnected && nodeConnectionError ? (
                 <tr>
@@ -659,7 +707,7 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                 </th>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.connected)}:</td>
+                <th>{intl.formatMessage(messages.connected)}:</th>
                 <td className={this.getClass(isConnected)}>
                   {isConnected
                     ? intl.formatMessage(messages.statusOn)
@@ -667,7 +715,7 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                 </td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.synced)}:</td>
+                <th>{intl.formatMessage(messages.synced)}:</th>
                 <td className={this.getClass(isSynced)}>
                   {isSynced
                     ? intl.formatMessage(messages.statusOn)
@@ -675,25 +723,25 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                 </td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.syncPercentage)}:</td>
+                <th>{intl.formatMessage(messages.syncPercentage)}:</th>
                 <td>{syncPercentage.toFixed(2)}%</td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.networkBlockHeight)}:</td>
+                <th>{intl.formatMessage(messages.networkBlockHeight)}:</th>
                 <td>{networkBlockHeight}</td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.localBlockHeight)}:</td>
+                <th>{intl.formatMessage(messages.localBlockHeight)}:</th>
                 <td>{localBlockHeight}</td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.remainingUnsyncedBlocks)}:</td>
+                <th>{intl.formatMessage(messages.remainingUnsyncedBlocks)}:</th>
                 <td className={remainingUnsyncedBlocksClasses}>
                   {remainingUnsyncedBlocks >= 0 ? remainingUnsyncedBlocks : '-'}
                 </td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.latestLocalBlockAge)}:</td>
+                <th>{intl.formatMessage(messages.latestLocalBlockAge)}:</th>
                 <td className={latestLocalBlockAgeClasses}>
                   {latestLocalBlockTimestamp > 0
                     ? `${latestLocalBlockAge} ms`
@@ -701,7 +749,7 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                 </td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.latestNetworkBlockAge)}:</td>
+                <th>{intl.formatMessage(messages.latestNetworkBlockAge)}:</th>
                 <td className={latestNetworkBlockAgeClasses}>
                   {latestNetworkBlockTimestamp > 0
                     ? `${latestNetworkBlockAge} ms`
@@ -709,7 +757,7 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                 </td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.localTimeDifference)}:</td>
+                <th>{intl.formatMessage(messages.localTimeDifference)}:</th>
                 <td className={styles.localTimeDifferenceItem}>
                   <button
                     onClick={() => onForceCheckLocalTimeDifference()}
@@ -729,7 +777,7 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                 </td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.systemTimeCorrect)}:</td>
+                <th>{intl.formatMessage(messages.systemTimeCorrect)}:</th>
                 <td className={this.getClass(isSystemTimeCorrect)}>
                   {isSystemTimeCorrect
                     ? intl.formatMessage(messages.statusOn)
@@ -737,7 +785,7 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                 </td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.systemTimeIgnored)}:</td>
+                <th>{intl.formatMessage(messages.systemTimeIgnored)}:</th>
                 <td className={this.getClass(!isSystemTimeIgnored)}>
                   {isSystemTimeIgnored
                     ? intl.formatMessage(messages.statusOn)
@@ -745,13 +793,15 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                 </td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.checkingNodeTime)}:</td>
+                <th>{intl.formatMessage(messages.checkingNodeTime)}:</th>
                 <td>
                   {isForceCheckingNodeTime
                     ? intl.formatMessage(messages.statusOn)
                     : intl.formatMessage(messages.statusOff)}
                 </td>
               </tr>
+            </tbody>
+            <tbody>
               <tr>
                 <th colSpan={2}>
                   {intl.formatMessage(messages.cardanoNodeStatus)}
@@ -769,9 +819,9 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
               </tr>
               {cardanoNodeEkgLink ? (
                 <tr>
-                  <td>
+                  <th>
                     {intl.formatMessage(messages.cardanoNodeDiagnostics)}:
-                  </td>
+                  </th>
                   <td>
                     <button
                       className={styles.realTimeStatusBtn}
@@ -787,7 +837,7 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                 </tr>
               ) : null}
               <tr>
-                <td>{intl.formatMessage(messages.cardanoNodeState)}:</td>
+                <th>{intl.formatMessage(messages.cardanoNodeState)}:</th>
                 <td>
                   {upperFirst(
                     cardanoNodeState != null
@@ -799,7 +849,7 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                 </td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.cardanoNodeResponding)}:</td>
+                <th>{intl.formatMessage(messages.cardanoNodeResponding)}:</th>
                 <td className={this.getClass(isNodeResponding)}>
                   {isNodeResponding
                     ? intl.formatMessage(messages.statusOn)
@@ -807,7 +857,7 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                 </td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.cardanoNodeSubscribed)}:</td>
+                <th>{intl.formatMessage(messages.cardanoNodeSubscribed)}:</th>
                 <td className={this.getClass(isNodeSubscribed)}>
                   {isNodeSubscribed
                     ? intl.formatMessage(messages.statusOn)
@@ -815,7 +865,7 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                 </td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.cardanoNodeTimeCorrect)}:</td>
+                <th>{intl.formatMessage(messages.cardanoNodeTimeCorrect)}:</th>
                 <td className={this.getClass(isNodeTimeCorrect)}>
                   {isNodeTimeCorrect
                     ? intl.formatMessage(messages.statusOn)
@@ -823,7 +873,7 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                 </td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.cardanoNodeSyncing)}:</td>
+                <th>{intl.formatMessage(messages.cardanoNodeSyncing)}:</th>
                 <td className={this.getClass(isNodeSyncing)}>
                   {isNodeSyncing
                     ? intl.formatMessage(messages.statusOn)
@@ -831,7 +881,7 @@ export default class DaedalusDiagnostics extends Component<Props, State> {
                 </td>
               </tr>
               <tr>
-                <td>{intl.formatMessage(messages.cardanoNodeInSync)}:</td>
+                <th>{intl.formatMessage(messages.cardanoNodeInSync)}:</th>
                 <td className={this.getClass(isNodeInSync)}>
                   {isNodeInSync
                     ? intl.formatMessage(messages.statusOn)