+++ source/renderer/app/themes/daedalus/cardano.js
@@ -512,7 +512,7 @@ const CARDANO_THEME_CONFIG = {
     '--theme-node-update-button-text-color': '#fafbfc',
   },
   notification: {
-    '--theme-notification-message-background-color': 'rgba(44, 187, 105, 0.88)',
+    '--theme-notification-message-background-color': 'rgba(44, 187, 105, 0.95)',
     '--theme-notification-message-text-color': '#fafbfc',
     '--theme-legacy-badge-background-color': '#ab1700',
     '--theme-legacy-notification-background-color': '#ab2712',